+++ tests/test_document.py
@@ -174,7 +174,13 @@ class DescribeDocument:
         picture = document.add_picture(path, width, height)
 
         run_.add_picture.assert_called_once_with(
-            path, width, height, description=None, title=None, svg_fallback=None
+            path,
+            width,
+            height,
+            description=None,
+            title=None,
+            svg_fallback=None,
+            honor_exif_orientation=True,
         )
         assert picture is picture_